backend-services/product_service/requirements.txt
django>=5.2,<6.0
djangorestframework
django-cors-headers
pymysql
cryptography
python-dotenv
uvicorn[standard]
